webapp/lambda/requirements.txt
psycopg2-binary==2.9.7
pandas==2.1.1
numpy==1.24.3